create utils and templates files for reusability in message.ts and voice.ts
