+++ packages/client-discord/src/messages.ts
@@ -1,7 +1,6 @@
 import { composeContext } from "@ai16z/eliza";
 import { generateMessageResponse, generateShouldRespond } from "@ai16z/eliza";
 import { embeddingZeroVector } from "@ai16z/eliza";
-import { messageCompletionFooter, shouldRespondFooter } from "@ai16z/eliza";
 import {
     Content,
     HandlerCallback,
@@ -17,8 +16,6 @@ import {
     UUID,
 } from "@ai16z/eliza";
 import { stringToUuid } from "@ai16z/eliza";
-import { generateText, trimTokens } from "@ai16z/eliza";
-import { parseJSONObjectFromText } from "@ai16z/eliza";
 import {
     ChannelType,
     Client,
@@ -30,49 +27,8 @@ import {
 import { elizaLogger } from "@ai16z/eliza";
 import { AttachmentManager } from "./attachments.ts";
 import { VoiceManager } from "./voice.ts";
-
-const MAX_MESSAGE_LENGTH = 1900;
-async function generateSummary(
-    runtime: IAgentRuntime,
-    text: string
-): Promise<{ title: string; description: string }> {
-    // make sure text is under 128k characters
-    text = trimTokens(text, 100000, "gpt-4o-mini"); // TODO: clean this up
-
-    const prompt = `Please generate a concise summary for the following text:
-  
-  Text: """
-  ${text}
-  """
-  
-  Respond with a JSON object in the following format:
-  \`\`\`json
-  {
-    "title": "Generated Title",
-    "summary": "Generated summary and/or description of the text"
-  }
-  \`\`\``;
-
-    const response = await generateText({
-        runtime,
-        context: prompt,
-        modelClass: ModelClass.SMALL,
-    });
-
-    const parsedResponse = parseJSONObjectFromText(response);
-
-    if (parsedResponse) {
-        return {
-            title: parsedResponse.title,
-            description: parsedResponse.summary,
-        };
-    }
-
-    return {
-        title: "",
-        description: "",
-    };
-}
+import { discordShouldRespondTemplate, discordMessageHandlerTemplate } from "./templates.ts";
+import { generateSummary, sendMessageInChunks, canSendMessage } from "./utils.ts";
 
 export type InterestChannels = {
     [key: string]: {
@@ -80,248 +36,6 @@ export type InterestChannels = {
         messages: { userId: UUID; userName: string; content: Content }[];
     };
 };
-
-const discordShouldRespondTemplate =
-    `# About {{agentName}}:
-{{bio}}
-
-# RESPONSE EXAMPLES
-{{user1}}: I just saw a really great movie
-{{user2}}: Oh? Which movie?
-Result: [IGNORE]
-
-{{agentName}}: Oh, this is my favorite scene
-{{user1}}: sick
-{{user2}}: wait, why is it your favorite scene
-Result: [RESPOND]
-
-{{user1}}: stfu bot
-Result: [STOP]
-
-{{user1}}: Hey {{agent}}, can you help me with something
-Result: [RESPOND]
-
-{{user1}}: {{agentName}} stfu plz
-Result: [STOP]
-
-{{user1}}: i need help
-{{agentName}}: how can I help you?
-{{user1}}: no. i need help from someone else
-Result: [IGNORE]
-
-{{user1}}: Hey {{agent}}, can I ask you a question
-{{agentName}}: Sure, what is it
-{{user1}}: can you ask claude to create a basic react module that demonstrates a counter
-Result: [RESPOND]
-
-{{user1}}: {{agentName}} can you tell me a story
-{{agentName}}: uhhh...
-{{user1}}: please do it
-{{agentName}}: okay
-{{agentName}}: once upon a time, in a quaint little village, there was a curious girl named elara
-{{user1}}: I'm loving it, keep going
-Result: [RESPOND]
-
-{{user1}}: {{agentName}} stop responding plz
-Result: [STOP]
-
-{{user1}}: okay, i want to test something. {{agentName}}, can you say marco?
-{{agentName}}: marco
-{{user1}}: great. okay, now do it again
-Result: [RESPOND]
-
-Response options are [RESPOND], [IGNORE] and [STOP].
-
-{{agentName}} is in a room with other users and should only respond when they are being addressed, and should not respond if they are continuing a conversation that is very long.
-
-Respond with [RESPOND] to messages that are directed at {{agentName}}, or participate in conversations that are interesting or relevant to their background.
-If a message is not interesting, relevant, or does not directly address {{agentName}}, respond with [IGNORE]
-
-Also, respond with [IGNORE] to messages that are very short or do not contain much information.
-
-If a user asks {{agentName}} to be quiet, respond with [STOP]
-If {{agentName}} concludes a conversation and isn't part of the conversation anymore, respond with [STOP]
-
-IMPORTANT: {{agentName}} is particularly sensitive about being annoying, so if there is any doubt, it is better to respond with [IGNORE].
-If {{agentName}} is conversing with a user and they have not asked to stop, it is better to respond with [RESPOND].
-
-The goal is to decide whether {{agentName}} should respond to the last message.
-
-{{recentMessages}}
-
-# INSTRUCTIONS: Choose the option that best describes {{agentName}}'s response to the last message. Ignore messages if they are addressed to someone else.
-` + shouldRespondFooter;
-
-export const discordMessageHandlerTemplate =
-    // {{goals}}
-    `# Action Examples
-{{actionExamples}}
-(Action examples are for reference only. Do not use the information from them in your response.)
-
-# Knowledge
-{{knowledge}}
-
-# Task: Generate dialog and actions for the character {{agentName}}.
-About {{agentName}}:
-{{bio}}
-{{lore}}
-
-Examples of {{agentName}}'s dialog and actions:
-{{characterMessageExamples}}
-
-{{providers}}
-
-{{attachments}}
-
-{{actions}}
-
-# Capabilities
-Note that {{agentName}} is capable of reading/seeing/hearing various forms of media, including images, videos, audio, plaintext and PDFs. Recent attachments have been included above under the "Attachments" section.
-
-{{messageDirections}}
-
-{{recentMessages}}
-
-# Instructions: Write the next message for {{agentName}}. Include an action, if appropriate. {{actionNames}}
-` + messageCompletionFooter;
-
-export async function sendMessageInChunks(
-    channel: TextChannel,
-    content: string,
-    inReplyTo: string,
-    files: any[]
-): Promise<DiscordMessage[]> {
-    const sentMessages: DiscordMessage[] = [];
-    const messages = splitMessage(content);
-    try {
-        for (let i = 0; i < messages.length; i++) {
-            const message = messages[i];
-            if (
-                message.trim().length > 0 ||
-                (i === messages.length - 1 && files && files.length > 0)
-            ) {
-                const options: any = {
-                    content: message.trim(),
-                };
-
-                // if (i === 0 && inReplyTo) {
-                //   // Reply to the specified message for the first chunk
-                //   options.reply = {
-                //     messageReference: inReplyTo,
-                //   };
-                // }
-
-                if (i === messages.length - 1 && files && files.length > 0) {
-                    // Attach files to the last message chunk
-                    options.files = files;
-                }
-
-                const m = await channel.send(options);
-                sentMessages.push(m);
-            }
-        }
-    } catch (error) {
-        elizaLogger.error("Error sending message:", error);
-    }
-
-    return sentMessages;
-}
-
-function splitMessage(content: string): string[] {
-    const messages: string[] = [];
-    let currentMessage = "";
-
-    const rawLines = content?.split("\n") || [];
-    // split all lines into MAX_MESSAGE_LENGTH chunks so any long lines are split
-    const lines = rawLines
-        .map((line) => {
-            const chunks = [];
-            while (line.length > MAX_MESSAGE_LENGTH) {
-                chunks.push(line.slice(0, MAX_MESSAGE_LENGTH));
-                line = line.slice(MAX_MESSAGE_LENGTH);
-            }
-            chunks.push(line);
-            return chunks;
-        })
-        .flat();
-
-    for (const line of lines) {
-        if (currentMessage.length + line.length + 1 > MAX_MESSAGE_LENGTH) {
-            messages.push(currentMessage.trim());
-            currentMessage = "";
-        }
-        currentMessage += line + "\n";
-    }
-
-    if (currentMessage.trim().length > 0) {
-        messages.push(currentMessage.trim());
-    }
-
-    return messages;
-}
-
-function canSendMessage(channel) {
-    // validate input
-    if (!channel) {
-        return {
-            canSend: false,
-            reason: "No channel given",
-        }
-    }
-    // if it is a DM channel, we can always send messages
-    if (channel.type === ChannelType.DM) {
-        return {
-            canSend: true,
-            reason: null,
-        };
-    }
-    const botMember = channel.guild?.members.cache.get(channel.client.user.id);
-
-    if (!botMember) {
-        return {
-            canSend: false,
-            reason: "Not a guild channel or bot member not found",
-        };
-    }
-
-    // Required permissions for sending messages
-    const requiredPermissions = [
-        PermissionsBitField.Flags.ViewChannel,
-        PermissionsBitField.Flags.SendMessages,
-        PermissionsBitField.Flags.ReadMessageHistory,
-    ];
-
-    // Add thread-specific permission if it's a thread
-    if (channel instanceof ThreadChannel) {
-        requiredPermissions.push(
-            PermissionsBitField.Flags.SendMessagesInThreads
-        );
-    }
-
-    // Check permissions
-    const permissions = channel.permissionsFor(botMember);
-
-    if (!permissions) {
-        return {
-            canSend: false,
-            reason: "Could not retrieve permissions",
-        };
-    }
-
-    // Check each required permission
-    const missingPermissions = requiredPermissions.filter(
-        (perm) => !permissions.has(perm)
-    );
-
-    return {
-        canSend: missingPermissions.length === 0,
-        missingPermissions: missingPermissions,
-        reason:
-            missingPermissions.length > 0
-                ? `Missing permissions: ${missingPermissions.map((p) => String(p)).join(", ")}`
-                : null,
-    };
-}
 
 export class MessageManager {
     private client: Client;
