+++ packages/client-discord/src/voice.ts
@@ -48,32 +48,7 @@ import {
     discordVoiceHandlerTemplate,
 } from "./templates.ts";
 import debounce from "lodash/debounce.js";
-
-export function getWavHeader(
-    audioLength: number,
-    sampleRate: number,
-    channelCount: number = 1,
-    bitsPerSample: number = 16
-): Buffer {
-    const wavHeader = Buffer.alloc(44);
-    wavHeader.write("RIFF", 0);
-    wavHeader.writeUInt32LE(36 + audioLength, 4); // Length of entire file in bytes minus 8
-    wavHeader.write("WAVE", 8);
-    wavHeader.write("fmt ", 12);
-    wavHeader.writeUInt32LE(16, 16); // Length of format data
-    wavHeader.writeUInt16LE(1, 20); // Type of format (1 is PCM)
-    wavHeader.writeUInt16LE(channelCount, 22); // Number of channels
-    wavHeader.writeUInt32LE(sampleRate, 24); // Sample rate
-    wavHeader.writeUInt32LE(
-        (sampleRate * bitsPerSample * channelCount) / 8,
-        28
-    ); // Byte rate
-    wavHeader.writeUInt16LE((bitsPerSample * channelCount) / 8, 32); // Block align ((BitsPerSample * Channels) / 8)
-    wavHeader.writeUInt16LE(bitsPerSample, 34); // Bits per sample
-    wavHeader.write("data", 36); // Data chunk header
-    wavHeader.writeUInt32LE(audioLength, 40); // Data chunk size
-    return wavHeader;
-}
+import { getWavHeader } from "./utils.ts";
 
 // These values are chosen for compatibility with picovoice components
 const DECODE_FRAME_SIZE = 1024;
